Content arrays are executed sequentially rather than in parallel
Previously, a content array was evaluated using Q.all, which made no guarantee about ordering.

Now that a content array can contain functions, this order matters. If a function in an array mutates some state, and a later function in that array accesses it, we need to make sure that works as expected.

This feels a bit hacky, but gets the job done.